+++ src/routes/TabbedRoutes.module.css
@@ -1,0 +1,22 @@
+.tabs-content {
+  display: flex;
+}
+
+.routerOutletLeftPane {
+  /* width: 30%; */
+}
+
+.routerOutletContents {
+  display: flex;
+  width: 100%;
+  height: 100%;
+
+  > * {
+    position: initial;
+    flex: 1;
+  }
+
+  > *:first-child {
+    max-width: 520px;
+  }
+}
