+++ src/routes/pages/posts/PostPage.tsx
@@ -50,7 +50,7 @@ export default function PostPage() {
   );
 }
 
-function PostPageContent({
+export function PostPageContent({
   id,
   commentPath,
   community,
